chore: move pkg fixtures
