+++ test/config/fixtures/pkg-zero-conf-noname/package.json
