+++ test/config/fixtures/pkg-zero-conf-nsname/index.js
@@ -4,7 +4,7 @@
  * compliance with the BSD 2-Clause License.
  */
 
-const apm = require('../../..').start({
+const apm = require('../../../..').start({
   disableSend: true,
   logLevel: 'off',
 });
