+++ test/config/fixtures/pkg-zero-conf-nsname/package.json
