+++ test/config/fixtures/pkg-zero-conf-sanitize/package.json
