+++ test/config/fixtures/pkg-zero-conf-valid/package.json
