+++ test/config/fixtures/pkg-zero-conf-weird/index.js
@@ -4,8 +4,7 @@
  * compliance with the BSD 2-Clause License.
  */
 
-const apm = require('../../..').start({
+const apm = require('../../../..').start({
   disableSend: true,
-  logLevel: 'off',
 });
 console.log(JSON.stringify(apm._conf));
