+++ test/config/fixtures/pkg-zero-conf-weird/package.json
